Refactor export endpoints into shared route.
