+++ server/api/org/org.controller.ts
@@ -1,8 +1,5 @@
-import { SearchResponse } from 'elasticsearch';
 import { Response } from 'express';
 import { Like } from 'typeorm';
-import { json2csvAsync } from 'json-2-csv';
-import moment from 'moment';
 import {
   ApiRequest,
   OrgParam,
@@ -11,11 +8,8 @@ import {
 import { Org } from './org.model';
 import {
   BadRequestError,
-  ForbiddenError,
-  InternalServerError,
   NotFoundError,
 } from '../../util/error-types';
-import elasticsearch from '../../elasticsearch/elasticsearch';
 
 class OrgController {
 
@@ -92,82 +86,6 @@ class OrgController {
     res.json(updatedOrg);
   }
 
-  async export(req: ApiRequest<null, null, ExportQuery>, res: Response) {
-    if (!req.appWorkspace) {
-      throw new ForbiddenError('You must be assigned to a workspace to export data.');
-    }
-
-    const startDate = moment(req.query.startDate ?? 0);
-    const endDate = moment(req.query.endDate);
-
-    //
-    // Get ES data and stream to client.
-    //
-    const index = req.appRole!.getKibanaIndex();
-    const scrollQueue = [] as SearchResponse<any>[];
-    try {
-      scrollQueue.push(await elasticsearch.search({
-        index,
-        size: 10000,
-        body: {
-          query: {
-            bool: {
-              must: [{
-                range: {
-                  Timestamp: {
-                    gte: startDate.toISOString(),
-                    lt: endDate.toISOString(),
-                  },
-                },
-              }],
-            },
-          },
-        },
-        scroll: '30s', // Keep the results "scrollable" for 30 seconds (this will be reset on every scroll).
-      }));
-    } catch (err) {
-      throw new InternalServerError(`Elasticsearch: ${err.message}`);
-    }
-
-    // Keep scrolling until we've sent all of the hits.
-    let hitsSent = 0;
-    while (scrollQueue.length) {
-      const chunk = scrollQueue.shift()!;
-
-      // Send the total hits before the first chunk so the client can calculate progress.
-      const isFirstChunk = (hitsSent === 0);
-      if (isFirstChunk) {
-        res.header('total-hits', `${chunk.hits.total}`);
-      }
-
-      const sourceArray = chunk.hits.hits.map(hit => hit._source);
-
-      const csvChunk = await json2csvAsync(sourceArray, {
-        prependHeader: isFirstChunk,
-        emptyFieldValue: '',
-      });
-
-      res.write(Buffer.from(csvChunk));
-
-      hitsSent += chunk.hits.hits.length;
-      if (hitsSent >= chunk.hits.total) {
-        break;
-      }
-
-      // Get the next response.
-      try {
-        scrollQueue.push(await elasticsearch.scroll({
-          scrollId: chunk._scroll_id!,
-          scroll: '30s',
-        }));
-      } catch (err) {
-        throw new Error(`Elasticsearch: ${err.message}`);
-      }
-    }
-
-    res.end();
-  }
-
 }
 
 type AddOrgBody = {
@@ -176,10 +94,5 @@ type AddOrgBody = {
 };
 
 type UpdateOrgBody = AddOrgBody;
-
-type ExportQuery = {
-  startDate?: string
-  endDate?: string
-};
 
 export default new OrgController();
