+++ server/api/roster/custom-roster-column.model.ts
@@ -2,7 +2,7 @@ import {
   Entity, Column, BaseEntity, JoinColumn, ManyToOne, PrimaryColumn,
 } from 'typeorm';
 import { Org } from '../org/org.model';
-import { RosterColumnType } from './roster.model';
+import { RosterColumnType } from './roster.types';
 
 @Entity()
 export class CustomRosterColumn extends BaseEntity {
