+++ server/api/roster/roster.model.ts
@@ -1,7 +1,21 @@
 import {
-  Entity, Column, BaseEntity, ManyToOne, CreateDateColumn, PrimaryGeneratedColumn,
+  BaseEntity,
+  Column,
+  CreateDateColumn,
+  Entity,
+  ManyToOne,
+  PrimaryGeneratedColumn,
 } from 'typeorm';
+import { snakeCase } from 'typeorm/util/StringUtils';
+import {
+  CustomColumns,
+  RosterColumnInfo,
+  RosterColumnType,
+} from './roster.types';
+import { Org } from '../org/org.model';
+import { Role } from '../role/role.model';
 import { Unit } from '../unit/unit.model';
+import { CustomRosterColumn } from './custom-roster-column.model';
 
 @Entity()
 export class Roster extends BaseEntity {
@@ -68,31 +82,79 @@ export class Roster extends BaseEntity {
 
     return Reflect.get(this, column.name) || null;
   }
-}
 
-export interface CustomColumns {
-  [key: string]: CustomColumnValue
-}
+  static getColumnSelect(column: RosterColumnInfo) {
+    // Make sure custom columns are converted to appropriate types
+    if (column.custom) {
+      switch (column.type) {
+        case RosterColumnType.Boolean:
+          return `(roster.custom_columns ->> '${column.name}')::BOOLEAN`;
+        case RosterColumnType.Number:
+          return `(roster.custom_columns ->> '${column.name}')::DOUBLE PRECISION`;
+        default:
+          return `roster.custom_columns ->> '${column.name}'`;
+      }
+    }
+    return `roster.${snakeCase(column.name)}`;
+  }
 
-export type CustomColumnValue = string | boolean | number | null;
+  static async queryAllowedRoster(org: Org, role: Role) {
+    //
+    // Query the roster, returning only columns and rows that are allowed for the role of the requester.
+    //
+    const columns = await Roster.getAllowedColumns(org, role);
+    const queryBuilder = Roster.createQueryBuilder('roster').select([]);
+    queryBuilder.leftJoin('roster.unit', 'u');
+    // Always select the id column
+    queryBuilder.addSelect('roster.id', 'id');
+    queryBuilder.addSelect('u.id', 'unit');
 
-export enum RosterColumnType {
-  String = 'string',
-  Boolean = 'boolean',
-  Date = 'date',
-  DateTime = 'datetime',
-  Number = 'number',
-}
+    // Add all columns that are allowed by the user's role
+    columns.forEach(column => {
+      queryBuilder.addSelect(Roster.getColumnSelect(column), column.name);
+    });
 
-export interface RosterColumnInfo {
-  name: string,
-  displayName: string,
-  type: RosterColumnType,
-  pii: boolean,
-  phi: boolean,
-  custom: boolean,
-  required: boolean,
-  updatable: boolean,
+    // Filter out roster entries that are not on the active roster or are not allowed by the role's index prefix.
+    return queryBuilder
+      .where('u.org_id = :orgId', { orgId: org.id })
+      .andWhere('(roster.end_date IS NULL OR roster.end_date >= CURRENT_DATE)')
+      .andWhere('(roster.start_date IS NULL OR roster.start_date <= CURRENT_DATE)')
+      .andWhere('u.id like :name', { name: role.indexPrefix.replace('*', '%') });
+  }
+
+  static async getAllowedColumns(org: Org, role: Role) {
+    const allColumns = await Roster.getColumns(org.id);
+    const fineGrained = !(role.allowedRosterColumns.length === 1 && role.allowedRosterColumns[0] === '*');
+    return allColumns.filter(column => {
+      let allowed = true;
+      if (fineGrained && role.allowedRosterColumns.indexOf(column.name) < 0) {
+        allowed = false;
+      } else if (!role.canViewPII && column.pii) {
+        allowed = false;
+      } else if (!role.canViewPHI && column.phi) {
+        allowed = false;
+      }
+      return allowed;
+    });
+  }
+
+  static async getColumns(orgId: number) {
+    const customColumns = (await CustomRosterColumn.find({
+      where: {
+        org: orgId,
+      },
+    })).map(customColumn => {
+      const columnInfo: RosterColumnInfo = {
+        ...customColumn,
+        displayName: customColumn.display,
+        custom: true,
+        updatable: true,
+      };
+      return columnInfo;
+    });
+    return [...baseRosterColumns, ...customColumns];
+  }
+
 }
 
 export const baseRosterColumns: RosterColumnInfo[] = [
